Remove backend validator tooling
