+++ packages/core/Cargo.toml
@@ -23,6 +23,4 @@ home = "0.5.11"
 hex = "0.4.3"
 rand = "0.9.2"
 sysinfo = "0.37.1"
-sha2 = "0.10.9"
 tempfile = "3.23.0"
-blst = "0.3.16"
