+++ packages/core/src/api/types.rs
@@ -1,7 +1,7 @@
+pub use crate::api::validator::{DepositData, ValidatorKey};
 pub use crate::domain::config::Config;
 pub use crate::domain::logs::LogsQuery;
 pub use crate::domain::operational_state::{OperationalMode, OperationalState};
 pub use crate::domain::package::{Package, PackageConfig, PackageRuntimeState};
 pub use crate::domain::system_info::SystemInfo;
-pub use crate::domain::validator::{DepositData, ValidatorKey};
 pub use crate::domain::web_service::WebServiceStatus;
