--- packages/core/src/application/validator/mod.rs
@@ -1,7 +1,0 @@
-mod create_deposit_data;
-mod filesystem;
-mod generate_keys;
-pub mod ports;
-
-pub use create_deposit_data::{CreateDepositDataParams, create_deposit_data};
-pub use generate_keys::{GenerateKeysParams, generate_keys};
