+++ packages/core/src/infra/mod.rs
@@ -3,5 +3,4 @@ pub mod docker;
 pub mod file;
 pub mod package;
 pub mod package_config;
-pub mod validator;
 pub mod web_service;
